requirements.txt
uvicorn
aiofiles
requests
pyyaml
opencv-python-headless
schedule 